rename graphiqlRelay -> explorerRelay
